requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.6
django-braces==1.15.0
django-ckeditor==6.5.1
django-filter==23.2
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-pwa==1.1.0
django-recaptcha==4.0.0
django-render-partial==0.4
django-social-share==2.3.0
django-star-ratings==0.9.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dnspython==2.3.0
fonttools==4.39.0
html5lib==1.1
idna==3.4
jdatetime==4.1.0
lxml==4.9.2
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pylokit==0.8.1
pypdf==3.5.2
pyphen==0.13.2
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.9.0
soupsieve==2.5
sqlparse==0.4.3
svglib==1.5.1
swapper==1.3.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
zopfli==0.2.2
